+++ example/src/forms/BasicForm/BasicForm.js
@@ -10,6 +10,7 @@ import Form, {
   Switch,
   TimePicker,
   DatePicker,
+  useFormState,
 } from '@detools/vue-form'
 
 export default {